Amend cert import and keychain setup for package signing
This commit mixes additions to the codesign flow such as
importing installer certs AND changes to how the methods are
called. Especially important is making sure to import the certs
with the -A flag. Otherwise, package signing hangs. This is
insecure—that is mitigated by the keychain being deleted
after notarization and then the VM being torn down.