Use std::asinh instead of our own calculation
I am pretty sure I was correctly calculating it. However,
it didn't sound very good while I was tweaking the condition
for when the normal calculation would be used. I think we
want that value as low as possible, because running the non
interpolated version is meant to be a fallback for when the
system would otherwise be in a bad condition (divide by 0).