Lower the tolerance governing when conventional asinh is used
As mentioned in the previous commit, I think we want to use the
interpolated calculation whenever possible. However, the asinh
sounds kind of wierd for values actually close to float epsilon
(around 1e-7). Changing to the std version helped a bit.

It also helps to not allow gain values smaller than 1 for the
saturator.