Lower minimum saturation gain
A lot of the excess saturation we were getting actually comes from the
main (asinh) waveshaper, which is still enabled at "0%". I don't want
to turn it off entirely at this setting, so now the lowest setting
corresponds to lower the gain slightly before the waveshaper, compensating
for it afterwards.