Clear clipper state on re-enable
I'm not sure if this is the right move, but it feels right.
We should probably clear state at some point, and I think
it's most straightforward to do so right before processing
again, as that's when the state would be wrong.

This is an effort to get rid of clicks, which still occur a little
bit. The solution here is simple in concept but will be a project
to implement: bake a dry/wet into the clipper processing, such that
turning it on or off with cause a fade up or down in that dry wet
mix. Will be a fun thing to implement in processors some day.