Change delay compensation according to output clip enablement
This required, among other things, using JUCE delay lines since
they had nice implementations for fractional delay that changes
during processing.

The overall approach is as follows: get the maximum delay in the
system, and round that up to the nearest int. Report that to host
and delay the unprocessed signal by that amound. Then, during processing,
delay the processed signal to have that same amount of latency, based
on a starting latency that changes depending on whether the clipper
is enabled.