Add a field for specifying Label Slider label text on construction
I've snuck in a change to our slider label text case as well.
In order to test the change and have a commit that builds, I
had to amend my use of this widget to the new interface.
Why use text I am about to change?

I could have defined an overload to prevent an interface break
but it all seemed like a fuss. Even this comment is a bit much.