Use radius instead of rw when defining sliderEdgeOffset
A sliderEdgeOffset greater than radius will cause the pointer
to draw incorrectly. Using radius instead of radius * 2, which
is how we get rw, makes the reasonable bounds of the offset
more clear.

We could clip it at sliderEdgeOffset > radius, but that seems
excessive, as the condition could only obtain due to coder error.